requirements.txt
streamlit
folium
streamlit-folium
scikit-learn
pandas
joblib
reportlab
requests
python-dotenv
